修复一些问题

+++ flowjson/jsoninter.cpp
@@ -173,8 +173,8 @@ bool JsonInter::getRoomInFile(QString fillname)
                 roomInfo* oneRoonInfo=new roomInfo;
 
                 QJsonArray coordinate_data=oneRoomObj.value("coordinate_data").toArray();
-                oneRoonInfo->roomPoint.setX(coordinate_data[0].toInt());
-                oneRoonInfo->roomPoint.setY(coordinate_data[1].toInt());
+                oneRoonInfo->roomPoint.setX(coordinate_data[0].toDouble());
+                oneRoonInfo->roomPoint.setY(coordinate_data[1].toDouble());
 
                 QJsonArray child_items=oneRoomObj.value("child_items").toArray();
                 foreach(auto oneitem,child_items)
